change price to double

+++ models/bookings.js
@@ -9,7 +9,7 @@ module.exports = (sequelize, DataTypes) => {
     invoiceNumber: DataTypes.STRING,
     note: DataTypes.TEXT,
     total: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DOUBLE(12,6),
       allowNull: false,
     },
     startDate: {
